Restrict paddle to game stage
Empirically determine the max X positions that the paddle can have and
still remain within the stage. This will require changing if the walls
or stage size changes.

X position is always changed on left or right arrow press. We just limit
what the maximum values can be. Perhaps we could optimize it so we don't
have to set the position if we are already at the max. This would save
us a call to new.