Make paddle affect ball direction
Based off the position the ball hits the paddle, calculate the new direction. The further from the center of the paddle, the sharper the angle. Effectively, when the ball hits the left side of the paddle, the ball moves left. When the ball hits the right side, it moves right.

This does not take into account the velocity of the ball, only its position when it hits the paddle.