Spawn bricks at their top left position
Brick sprites are positioned around their center. But for the config file coordinates, we want to position the brick sprites around their top left corner. So do a little extra math to position the brick using the top left corner as the pivot. It's not pixel perfect but it looks alright.