Apply size defined in the bricks.cfg file to the bricks
* Cache the SpriteRenderer of the brick prefab
Because we call GetComponent() way too often. Better to cache this.

The sizes defined in the bricks.cfg file are an integer multiplier of the default size. Which is roughly 0.5 x 0.5 game pixels, or also 1x1 cfg file pixels. So a size of 2,2 defined in the bricks.cfg file would produce a brick of size 1x1 game pixels.