Allow the game to end
Add hashSet to track all the bricks in the game. This is an efficient way to insert, search, and remove bricks.

When all the bricks are destroyed, displays a "Game Over!" text to the console log.

Use the event delegate messaging system to efficiently communicate between bricks and the brick manager, and reduce coupling.