Give balls a default state of 'stopped'
Give balls a default state of 'stopped', which means they will not move. Only when the caller explicitly tells the ball to move will the ball start moving. Change the code so that the functionality remains unchanged yet uses this new state paradigm.

The thinking behind this change is that I am going to implement a hack soon to nudge the ball's y velocity when it approaches 0. This way we won't be stuck in a position where the ball bounces along the x axis forever. But we don't want this hack to always take effect when the y velocity is zero. Only when the ball is in a state of motion.