Restart the game on "Play again" button press
This is actually a relatively large change. The biggest hurdle was to destroy all the current balls. As such, changed the ball management system. Instead of counting how many balls are in play, we store each ball GameObject in a HashSet. This required updating the code in various areas to add and remove from this set of balls as necessary.