Change the camera size based on the screen resolution
Use formula from this forum thread on programmatically calcuating the
correct camera size based on the screen resolution and our desired
aspect ratio:
http://forum.unity3d.com/threads/android-resolutions.211905/

The camera now properly zooms out on Nexus 7 portrait mode to display
the entire game scene. Although as a consequence the entire bounding box
is now visible, so a bit ugly.

We setup the camera at game init as well as whenever the device rotates
(only applies to mobile).