Modify scene walls to form a nearly perfect rectangle
Simply for aesthetics purposes. Adjust the walls of the scene so they
form a more complete rectangle. This is important now because if the
screen resolution is too small, then the player will see the entire
scene. So at the very least try to make the scene more aesthetically
pleasing.