Ignore collisions between balls
Add a new layer "Ignore Ball Collisions" to the ball prefab.
This change ignores collisions between ball objects.

I find that this fits better with my expectation of the game physics.
Before, when ball collided with one another, there was a good chance
that one of the balls would lose most of its velocity, giving the other
ball most of its velocity. Now that balls no longer impact one another,
all the balls are moving at a fast speed and gives the game a sense of
excitement.