Change ball release from double tap to single tap
Single tap is the more intuitive approach to releasing the ball from the
paddle. High level implementation details: if the start and end touch
coordinates are roughly the same then it is a tap. However, touch
movement or GUI overlay on start touch do not count. If we include the
GUI overlay as a touch, then when the user hits the start button, the
ball will immediately start moving.

Also fix a small bug where we forget to reset the game start status when
the game is over. Now whenver a new game is started, the new game will
be registered.